Move into FlowFuse Category, change branding & node name in prep for further changes re: db creds
