+++ .github/workflows/project-automation.yml
@@ -1,0 +1,10 @@
+name: Project Automation
+on:
+  issues:
+    types:
+      - opened
+jobs:
+  add_to_product_board:
+    uses: flowfuse/.github/.github/workflows/project-automation.yml@main
+    secrets:
+      token: ${{ secrets.PROJECT_ACCESS_TOKEN }}
