+++ nodes/locales/en-US/query-config.json
@@ -1,7 +1,6 @@
 {
-	"postgresql": {
+	"tables-query-config": {
 		"label": {
-			"name": "Name",
 			"host": "Host",
 			"port": "Port",
 			"database": "Database",
@@ -11,28 +10,24 @@
 			"applicationName": "Application name",
 			"max": "Maximum size",
 			"idle": "Idle Timeout",
-			"connectionTimeout": "Connection Timeout",
-			"server": "Server",
-			"query": "Query",
-			"split": "Split results in multiple messages",
-			"rowsPerMsg": "Number of rows per message"
+			"connectionTimeout": "Connection Timeout"
+		},
+		"tab": {
+			"connection": "Connection",
+			"security": "Security",
+			"pool": "Pool"
 		},
 		"placeholder": {
 			"name": "dbConnection",
 			"host": "127.0.0.1",
 			"port": "5432",
 			"database": "dbExample",
-			"user": "dbUser",
-			"password": "dbPassword",
 			"applicationName": "",
 			"max": "10",
 			"idle": "1000 (Milliseconds)",
-			"connectionTimeout": "10000 (Milliseconds)"
-		},
-		"tab": {
-			"connection": "Connection",
-			"security": "Security",
-			"pool": "Pool"
+			"connectionTimeout": "10000 (Milliseconds)",
+			"user": "dbUser",
+			"password": "dbPassword"
 		},
 		"title": {
 			"applicationName": "The name of the application that created this Client instance.",
