+++ nodes/query.html
@@ -1,0 +1,109 @@
+<style>
+.nr-tables-query-node text {
+	/* fill: white; */
+}
+</style>
+
+<script type="text/javascript">
+	/* global RED:false, $:false */
+	RED.nodes.registerType('tables-query', {
+		category: 'FlowFuse',
+		color: '#5b85a7',
+		defaults: {
+			name: {
+				value: '',
+			},
+			query: {
+				value: 'SELECT * FROM ;',
+			},
+			queryConfig: {
+				type: 'tables-query-config',
+				required: true,
+			},
+			split: {
+				value: false,
+			},
+			rowsPerMsg: {
+				value: 1,
+			},
+			outputs: {
+				value: 1,
+			},
+		},
+		inputs: 1,
+		icon: 'flowfuse.svg',
+		align: 'left',
+		paletteLabel: 'query',
+		label: function () {
+			return this.name || 'query';
+		},
+		labelStyle: function () {
+			const classes = 'nr-tables-query-node';
+			return this.name ? `${classes} node_label_italic` : classes;
+		},
+		oneditprepare: function () {
+			$('#node-input-split').prop('checked', this.split);
+
+			// when node-input-split is modified, react:
+			$('#node-input-split').on('change', function () {
+				// if checked, show the rowsPerMsg input
+				if ($('#node-input-split').is(':checked')) {
+					$('#node-input-rowsPerMsg-container').show();
+				} else {
+					$('#node-input-rowsPerMsg-container').hide();
+				}
+			});
+
+			$('#node-input-rowsPerMsg').value = this.split ? this.rowsPerMsg : 1;
+			this.editor = RED.editor.createEditor({
+				id: 'node-input-editor',
+				mode: 'ace/mode/sql',
+				value: $('#node-input-query').val(),
+			});
+			this.editor.focus();
+		},
+		oneditsave: function () {
+			$('#node-input-query').val(this.editor.getValue());
+			delete this.editor;
+		},
+	});
+</script>
+
+<script type="text/x-red" data-template-name="tables-query">
+	<div class="form-row">
+		<label for="node-input-name">
+			<i class="icon-tag"></i>
+			<span data-i18n="tables-query.label.name"></span>
+		</label>
+		<input type="text" id="node-input-name" data-i18n="[placeholder]node-red:common.label.name" />
+	</div>
+	<div class="form-row">
+		<label for="node-input-queryConfig">
+			<i class="fa fa-server"></i>
+			<span data-i18n="tables-query.label.server"></span>
+		</label>
+		<input type="text" id="node-input-queryConfig" />
+	</div>
+	<div class="form-row">
+		<input type="checkbox" id="node-input-split" style="display: inline-block; width: auto; vertical-align: top;" />
+		<label for="node-input-split" style="width: auto;">
+			<span data-i18n="tables-query.label.split"></span>
+		</label>
+	</div>
+	<div class="form-row" id="node-input-rowsPerMsg-container">
+		<label for="node-input-rowsPerMsg">
+			<span data-i18n="tables-query.label.rowsPerMsg"></span>
+		</label>
+		<input type="number" id="node-input-rowsPerMsg" placeholder="1" value="1" min="1" />
+	</div>
+	<div class="form-row" style="position: relative; margin-bottom: 0px;">
+		<label for="node-input-query">
+			<i class="fa fa-file-code-o"></i>
+			<span data-i18n="tables-query.label.query"></span>
+		</label>
+		<input type="hidden" id="node-input-query" autofocus="autofocus" />
+	</div>
+	<div class="form-row node-text-editor-row">
+		<div style="height: 300px; min-height: 150px;" class="node-text-editor" id="node-input-editor"></div>
+	</div>
+</script>
