+++ package.json
@@ -1,12 +1,17 @@
 {
-	"name": "node-red-contrib-postgresql",
-	"version": "0.15.1",
-	"description": "Node-RED node for PostgreSQL, supporting parameters, split, back-pressure",
+	"name": "@flowfuse/nr-tables-nodes",
+	"version": "0.1.0",
+	"description": "Nodes for use with the FlowFuse Tables offering, allowing developers to write and run queries against databases inside FlowFuse Tables.",
+	"homepage": "https://github.com/FlowFuse/nr-tables-nodes#readme",
 	"author": {
-		"name": "Alexandre Alapetite",
-		"url": "https://github.com/Alkarex"
+		"name": "FlowFuse Inc.",
+		"url": "https://github.com/flowfuse"
 	},
 	"contributors": [
+		{
+			"name": "Alexandre Alapetite",
+			"url": "https://github.com/Alkarex"
+		},
 		{
 			"name": "Andrea Batazzi",
 			"url": "https://github.com/andreabat"
@@ -23,7 +28,10 @@
 	"license": "Apache-2.0",
 	"repository": {
 		"type": "git",
-		"url": "https://github.com/alexandrainst/node-red-contrib-postgresql"
+		"url": "https://github.com/FlowFuse/nr-tables-nodes"
+	},
+	"bugs": {
+		"url": "https://github.com/FlowFuse/nr-tables-nodes/issues"
 	},
 	"keywords": [
 		"backpressure",
@@ -32,15 +40,17 @@
 		"nodered",
 		"postgres",
 		"postgresql",
-		"timescale"
+		"timescale",
+		"flowfuse"
 	],
 	"engines": {
-		"node": ">=8"
+		"node": ">=16.x"
 	},
 	"node-red": {
-		"version": ">=0.20.0",
+		"version": ">=3.1.0",
 		"nodes": {
-			"postgresql": "postgresql.js"
+			"tables-query": "nodes/query.js",
+			"tables-query-config": "nodes/query-config.js"
 		}
 	},
 	"dependencies": {
